Centralize delete statement with connections
